Fixxed the Overshoot error in 2 Ways, Controller works..
Ive Fixed the overshooting error in 2 ways:
either use the inner if, or the outer if. both ways should work,
otherwise feel free to test it.

Somehow I made the shit function correctly. At least with the
Controller. Just Check wether it works with the sensor.. maybe the
sensor is the problem..